improve softkeyboard handling
When the softkeyboard has been opened at least once in the chat fragment,
it needs to be hidden on fragment change since the chat fragment is still
in a "visible" state and the EditText keeps focus. As of now this is done
by hiding the keyboard whenever a new tab is selected. Unfortunately this
doesn't handle the case when the Activity is resumed via notification
since the EditText still gets the focus and therefore the keyboard is
opened again.
This patch works around this by making the parent Layout focusable and
removing focus from the EditText when leaving the chat fragment. Also only
do this when we're leaving the chat fragment.