Don't query osrm api for sternfahrt routes
getRoad() will always query the osrm API for routing information. As this is
done on the main thread the app will be unresponsive. Additionally most queries
will fail because the generated URLs are too long. In consequence getRoad()
will fall back to generating the Road by only using the waypoints.
Since the raw waypoints are already precise enough and none of the additional
information is used, just construct the Road directly.