Fix Marker InfoWindow handling
By default omsbonuspack markers open an info bubble onClick. This requires a
"bonuspack_bubble.xml" layout to be present, which is inflated on marker
instantiation. We don't have that, so every marker throws an exception.
Since the infowindow is not needed, handle this by adding a dummy layout and
by calling setInfoWindow(null) on every marker.