make service a foreground service
This makes it less likely that the process is killed by android due to memory pressure.
This also moves all the notification handling to the service and and makes sure the
notification is shown for the entire service lifecycle.
Note that there is no need to explicitly cancel the notification as this is automatically
taken care of when the service stops.