Fix memory leak due to osmbonuspack bug
The Marker class holds a static reference to a DefaultInfoWindow instance
which in turn holds a reference to the MapView. Thus it leaks the entire
view hierarchy when the fragment is detached.
Work around this by sub classing marker and implementing Overlay's
onDetach() method to null out the static references there.
Also move common initialization code to the subclass.