Determine and save storage location for map cache
Up until now we relied on osmdroid's default behavior which uses
outdated and unreliable methods to determine all available storages and
selects the one with the most free space. This is done on every map
creation and thus can lead to multiple cache files if the free space
changes.
This implements our own logic to determine the location with the most
free space on first run. That location is then saved and used in future
runs.
For now this only checks the app private locations returned by
getFilesDir() and getExternalFilesDirs(). These don't require the
WRITE_EXTERNAL_STORAGE permission and are deleted on app uninstall.