Change logic for when to post user's location
Before this, if there were no location updates from the system, we
stopped sending the last location after 30 secs (see removed method
OwnLocationModel.isLocationFresh()). This was done to not send old
locations that came from getLastKnownLocation() or if the location
wasn't updated due to environment conditions or if the user disabled
location services while the app was running.
This had the side effect of also not posting the location if a user
doesn't move for example at the beginning or the end of the event.
In this case we also don't get location updates from the system
since we're under the distance threshold
(LocationUpdateManager.LOCATION_REFRESH_DISTANCE).
To alleviate this we now always post the location but only if
location services on the device are enabled. This might still send
outdated locations due to environment conditions but should be the
better compromise.