Update to targetSdk 35 and fix missing Notification
We need to urgently update to API 35, so the app will remain available on play store.
At the same time we had some issues on newer API Versions not showing the service
notification. This lead to people leaving tracking on unintentionally. This has been fixed.
To simplify permission handling, we handle location and notification access together.